feat: Adaptation logic

+++ store/src/main/java/org/apache/rocketmq/store/timer/rocksdb/TimerMessageKVStore.java
@@ -107,5 +107,5 @@ public interface TimerMessageKVStore {
      * @param columnFamily the column family of the timer message kv store.
      * @return the checkpoint of the timer message kv store.
      */
-    int getCheckpoint(byte[] columnFamily);
+    long getCheckpoint(byte[] columnFamily);
 }
